blocking monitor working to the ui

+++ QIQO.Monitor.SQLServer.Data/Entities/BlockingData.cs
@@ -4,13 +4,13 @@
     {
         public string LockType { get; set; }
         public string Database { get; set; }
-        public int BlockObject { get; set; }
+        public long BlockObject { get; set; }
         public string LockRequest { get; set; }
         public int WaiterSid { get; set; }
-        public int WaitTime { get; set; }
+        public long WaitTime { get; set; }
         public string WaiterBatch { get; set; }
         public string WaiterStatement { get; set; }
-        public int BlockerSid { get; set; }
+        public short BlockerSid { get; set; }
         public string BlockerBatch { get; set; }
     }
 }
